\name{ChangePoints}
\alias{ChangePoints}
\title{Change Points of a \code{numeric} vector}
\usage{
ChangePoints(x)
}
\arguments{
  \item{x}{\code{numeric} vector.}
}
\value{
The vector of index of change points.
}
\description{
Finds changes points in the \code{numeric} vector.
}
\details{
The end points of the vector will be always included in the
results.
}

