\name{plot.pvar}
\alias{plot.pvar}
\title{plot.pvar}
\usage{
\method{plot}{pvar}(x, main = "p-variation", ylab = x$dname, sub = "p="
  \%.\% round(x$p, 5) \%.\% ", p-variation: " \%.\% formatC(x$value, 5, format =
  "f"), col.PP = 2, cex.PP = 0.5, ...)
}
\arguments{
  \item{x}{asd}

  \item{main}{asd}

  \item{ylab}{asd}

  \item{sub}{asd}

  \item{col.PP}{asd}

  \item{cex.PP}{asd}

  \item{\dots}{asd}
}
\description{
plot.pvar
}

