% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddPvarC}
\alias{AddPvarC}
\title{Addition of p-variation (in C++)}
\usage{
AddPvarC(PV1, PV2, AddIfPossible = TRUE)
}
\arguments{
\item{PV1}{an object of the class \code{pvar}.}

\item{PV2}{an object of the class \code{pvar}, which has the same \code{p} value as PV1 object.}

\item{AddIfPossible}{\code{logical}. If TRUE (the default), then is is assumed, that two samples has common point. So,
the end of PV1 and the begging of PV2 will be treated as one point if it has the same value.}
}
\value{
An object of the class \code{pvar}.
}
\description{
An internal function(written in C++) that merges two objects of pvar and effectively recalculates the p-variation of joined sample.
}
\details{
This is an internal function, therefore, users should 
not call this function directly (rather use \code{\link{AddPvar}} or \code{pv1 + pv2}).
}
\keyword{internal}
