% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-functions.R
\name{\%.\%}
\alias{\%.\%}
\title{Concatenate strings}
\usage{
x \%.\% y
}
\arguments{
\item{x}{asd}

\item{y}{asd}
}
\value{
A character string of the concatenated values.
}
\description{
Concatenate Strings
}
\details{
The same result may be achieved with \code{paste}, but in some circumstance this function is more user friendly.
}
\examples{
paste('I ', 'love ', 'R.', sep='')
'I ' \%.\% 'love ' \%.\% 'R.'

x = c(2,1,6,7,9)
paste('The length of vector (', paste(x , sep='', collapse =','), ') is ', length(x) , sep='')
'The length of vector (' \%.\% paste(x , sep='', collapse =',') \%.\% ') is ' \%.\% length(x)
}
\seealso{
\code{\link{paste}}
}
