\docType{data}
\name{DataSets}
\alias{DataSets}
\alias{MeanCoef}
\alias{PvarQuantileDF}
\alias{SdCoef}
\title{CpG islands - human - genome build: GRCh37/hg19}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
UCSC Table Browser
}
\usage{
PvarQuantileDF
MeanCoef
SdCoef
}
\description{
This data set list the genomic locations of human CpG
islands, with coordinates based on the GRCh37 / hg19 genome
build.
}
\author{
Mark Cowley, 2012-03-05
}

